Fix validateIsLossless so the tests run.
The way validateIsLossless was being exposed led to its tests
never being run. It's a good example of why mutation is bad.
Here we reconfigure it to require it to be mixed into a mutable
specification, because that is the only composition mechanism
we have in the mutable specification scenario.